Move check for NaN values templates into header.
